switch to f64 in al model because perf on 32 bit is for cowards

+++ azur_lane/src/skill.rs
@@ -66,7 +66,7 @@ define_data_enum! {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct BuffInfo {
     pub buff_id: u32,
-    pub probability: f32,
+    pub probability: f64,
     #[serde(default, skip_serializing_if = "crate::data_def::is_default")]
     pub level: u32,
 }
