+++ azur_lane_data_collector/src/enhance/meta.rs
@@ -8,7 +8,7 @@ pub fn add_repair(ship: &mut ShipData, table: &LuaTable) -> LuaResult<()> {
     let effect: LuaTable = table.get("effect_attr")?;
     
     let attr: String = context!(effect.get(1); "repair's effect_attr name for meta ship id {}", ship.group_id)?;
-    let value: f32 = effect.get(2)?;
+    let value: f64 = effect.get(2)?;
 
     super::add_to_stats_base(&mut ship.stats, &attr, value);
 
@@ -22,7 +22,7 @@ pub fn add_repair_effect(ship: &mut ShipData, table: &LuaTable) -> LuaResult<()>
     let effect_attr: Vec<LuaTable> = table.get("effect_attr")?;
     for effect in effect_attr {
         let attr: String = context!(effect.get(1); "repair_effect's effect_attr name for meta ship id {}", ship.group_id)?;
-        let value: f32 = effect.get(2)?;
+        let value: f64 = effect.get(2)?;
 
         super::add_to_stats_base(&mut ship.stats, &attr, value);
     }
